+++ hprof-heap/src/main/java/org/netbeans/lib/profiler/heap/NumberList.java
@@ -1,48 +1,26 @@
 /*
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
  *
- * Copyright 1997-2010 Oracle and/or its affiliates. All rights reserved.
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Oracle and Java are registered trademarks of Oracle and/or its affiliates.
- * Other names may be trademarks of their respective owners.
- *
- * The contents of this file are subject to the terms of either the GNU
- * General Public License Version 2 only ("GPL") or the Common
- * Development and Distribution License("CDDL") (collectively, the
- * "License"). You may not use this file except in compliance with the
- * License. You can obtain a copy of the License at
- * http://www.netbeans.org/cddl-gplv2.html
- * or nbbuild/licenses/CDDL-GPL-2-CP. See the License for the
- * specific language governing permissions and limitations under the
- * License.  When distributing the software, include this License Header
- * Notice in each file and include the License file at
- * nbbuild/licenses/CDDL-GPL-2-CP.  Oracle designates this
- * particular file as subject to the "Classpath" exception as provided
- * by Oracle in the GPL Version 2 section of the License file that
- * accompanied this code. If applicable, add the following below the
- * License Header, with the fields enclosed by brackets [] replaced by
- * your own identifying information:
- * "Portions Copyrighted [year] [name of copyright owner]"
- *
- * Contributor(s):
- * The Original Software is NetBeans. The Initial Developer of the Original
- * Software is Sun Microsystems, Inc. Portions Copyright 1997-2006 Sun
- * Microsystems, Inc. All Rights Reserved.
- *
- * If you wish your version of this file to be governed by only the CDDL
- * or only the GPL Version 2, indicate your decision by adding
- * "[Contributor] elects to include this software in this distribution
- * under the [CDDL or GPL Version 2] license." If you do not indicate a
- * single choice of license, a recipient has the option to distribute
- * your version of this file under either the CDDL, the GPL Version 2 or
- * to extend the choice of license to its licensees as provided above.
- * However, if you add GPL Version 2 code and therefore, elected the GPL
- * Version 2 license, then the option applies only if the new code is
- * made subject to such option by the copyright holder.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 package org.netbeans.lib.profiler.heap;
 
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.RandomAccessFile;
@@ -54,6 +32,7 @@ import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 import java.util.Set;
 
 /**
@@ -66,27 +45,28 @@ class NumberList {
     private final RandomAccessFile data;
     private final int numberSize;
     private final int blockSize;
-    private final Map<Long/*offset*/, byte[]/*block*/> blockCache;
-    private final Set<Long> dirtyBlocks;
+	private final Map<Long/*offset*/, byte[]/*block*/> blockCache;
+	private final Set<Long> dirtyBlocks;
     private long blocks;
     private MappedByteBuffer buf;
     private long mappedSize;
-
-    NumberList(long dumpFileSize) throws IOException {
-        this(bytes(dumpFileSize));
+    private CacheDirectory cacheDirectory;
+    
+    NumberList(long dumpFileSize, CacheDirectory cacheDir) throws IOException {
+        this(bytes(dumpFileSize), cacheDir);
     }
-
-    NumberList(int elSize) throws IOException {
-        dataFile = File.createTempFile("NBProfiler", ".ref"); // NOI18N
+    
+    NumberList(int elSize, CacheDirectory cacheDir) throws IOException {
+        dataFile = cacheDir.createTempFile("NBProfiler", ".ref"); // NOI18N
         data = new RandomAccessFile(dataFile, "rw"); // NOI18N
         numberSize = elSize;
-        blockCache = new BlockLRUCache<Long, byte[]>();
+        blockCache = new BlockLRUCache();
         dirtyBlocks = new HashSet<Long>(100000);
         blockSize = (NUMBERS_IN_BLOCK + 1) * numberSize;
-        dataFile.deleteOnExit();
+        cacheDirectory = cacheDir;
         addBlock(); // first block is unused, since it starts at offset 0
     }
-
+     
     private static int bytes(long number) {
         if ((number & ~0xFFL) == 0L) {
             return 1;
@@ -111,12 +91,14 @@ class NumberList {
         }
         return 8;
     }
-
+    
     protected void finalize() throws Throwable {
-        dataFile.delete();
+        if (cacheDirectory.isTemporary()) {
+            dataFile.delete();
+        }
         super.finalize();
     }
-
+    
     long addNumber(long startOffset,long number) throws IOException {
         int slot;
         byte[] block = getBlock(startOffset);
@@ -136,7 +118,7 @@ class NumberList {
         writeNumber(nextBlock,block,0,number); // write number to first position in the new block
         return nextBlock;
     }
-
+    
     long addFirstNumber(long number1,long number2) throws IOException {
         long blockOffset = addBlock();
         byte[] block = getBlock(blockOffset);
@@ -144,7 +126,7 @@ class NumberList {
         writeNumber(blockOffset,block,1,number2);
         return blockOffset;
     }
-
+    
     void putFirst(long startOffset,long number) throws IOException {
         int slot;
         long offset = startOffset;
@@ -154,7 +136,7 @@ class NumberList {
             for (slot=0;slot<NUMBERS_IN_BLOCK;slot++) {
                 long el = readNumber(block,slot);
                 if (offset == startOffset && slot == 0) { // first block
-                    if (number == el) { // already first element
+                    if (number == el) { // already first element 
                         return;
                     }
                     movedNumber = el;
@@ -162,7 +144,7 @@ class NumberList {
                 } else if (el == 0L) { // end of the block, move to next one
                     break;
                 } else if (el == number) { // number is already in the list
-                    writeNumber(offset,block,slot,movedNumber);    // replace number and return
+                    writeNumber(offset,block,slot,movedNumber);    // replace number and return                
                     return;
                 }
             }
@@ -173,16 +155,20 @@ class NumberList {
             }
         }
     }
-
+    
     long getFirstNumber(long startOffset) throws IOException {
         byte[] block = getBlock(startOffset);
         return readNumber(block,0);
+    }
+    
+    LongIterator getNumbersIterator(long startOffset) throws IOException {
+        return new NumberIterator(startOffset);
     }
 
     List<Long> getNumbers(long startOffset) throws IOException {
         int slot;
         List<Long> numbers = new ArrayList<Long>();
-
+        
         for(;;) {
             byte[] block = getBlock(startOffset);
             for (slot=0;slot<NUMBERS_IN_BLOCK;slot++) {
@@ -199,18 +185,18 @@ class NumberList {
             startOffset = nextBlock;
         }
     }
-
+    
     private void mmapData() {
-        if (buf == null && blockSize*blocks < Integer.MAX_VALUE) {
+        if (buf == null) {
             try {
-                buf = data.getChannel().map(FileChannel.MapMode.READ_WRITE, 0, data.length());
-                mappedSize = blockSize*blocks;
+                mappedSize = Math.min(blockSize*blocks, Integer.MAX_VALUE-blockSize+1);
+                buf = data.getChannel().map(FileChannel.MapMode.READ_WRITE, 0, mappedSize);
             } catch (IOException ex) {
                 ex.printStackTrace();
             }
         }
     }
-
+    
     void flush() {
         try {
             flushDirtyBlocks();
@@ -220,18 +206,18 @@ class NumberList {
             ex.printStackTrace();
         }
     }
-
+    
     private long getOffsetToNextBlock(byte[] block) {
         return readNumber(block,NUMBERS_IN_BLOCK);
     }
-
+    
     private long readNumber(byte[] block,int slot) {
         int offset = slot*numberSize;
         long el = 0;
 //        for (int i=0;i<numberSize;i++) {
 //            el <<= 8;
 //            el |= ((int)block[offset+i]) & 0xFF;
-//        }
+//        }    
         if (numberSize == 4) {
             return ((long)getInt(block,offset)) & 0xFFFFFFFFL;
         } else if (numberSize == 8) {
@@ -259,16 +245,16 @@ class NumberList {
               ((buf[i++] & 255) <<  8) +
               ((buf[i++] & 255) <<  0));
     }
-
+        
     private void writeNumber(long blockOffset,byte[] block,int slot,long element) throws IOException {
         if (blockOffset < mappedSize) {
             long offset = blockOffset+slot*numberSize;
             buf.position((int)offset);
             for (int i=numberSize-1;i>=0;i--) {
                 byte el = (byte)(element >> (i*8));
-
+                
                 buf.put(el);
-            }
+            }            
         } else {
             Long offsetObj = new Long(blockOffset);
             int offset = slot*numberSize;
@@ -282,7 +268,7 @@ class NumberList {
             }
         }
     }
-
+    
     private byte[] getBlock(long offset) throws IOException {
         byte[] block;
         if (offset < mappedSize) {
@@ -315,8 +301,7 @@ class NumberList {
         if (dirtyBlocks.isEmpty()) {
             return;
         }
-        Long[] dirty=new Long[dirtyBlocks.size()];
-        dirtyBlocks.toArray(dirty);
+        Long[] dirty=(Long[])dirtyBlocks.toArray(new Long[0]);
         Arrays.sort(dirty);
         byte blocks[] = new byte[1024*blockSize];
         int dataOffset = 0;
@@ -333,7 +318,7 @@ class NumberList {
                 data.write(blocks,0,dataOffset);
                 dataOffset = 0;
                 System.arraycopy(block,0,blocks,dataOffset,blockSize);
-                dataOffset+=blockSize;
+                dataOffset+=blockSize;                
                 lastBlockOffset = blockOffset;
             }
         }
@@ -342,16 +327,98 @@ class NumberList {
         dirtyBlocks.clear();
     }
 
+    //---- Serialization support
+    void writeToStream(DataOutputStream out) throws IOException {
+        out.writeUTF(dataFile.getAbsolutePath());
+        out.writeInt(numberSize);
+        out.writeLong(blocks);
+        out.writeBoolean(buf != null);        
+    }
+
+    NumberList(DataInputStream dis, CacheDirectory cacheDir) throws IOException {
+        boolean mmaped;
+        
+        cacheDirectory = cacheDir;
+        dataFile = cacheDirectory.getCacheFile(dis.readUTF());
+        data = new RandomAccessFile(dataFile, "rw"); // NOI18N
+        numberSize = dis.readInt();
+        blocks = dis.readLong();
+        mmaped = dis.readBoolean();
+        blockCache = new BlockLRUCache();
+        dirtyBlocks = new HashSet<Long>(100000);
+        blockSize = (NUMBERS_IN_BLOCK + 1) * numberSize;
+        if (mmaped) {
+            mmapData();
+        }
+    }    
+    
+    private class NumberIterator extends LongIterator {
+        private int slot;
+        private byte[] block;
+        private long nextNumber;
+
+        private NumberIterator(long startOffset) throws IOException {
+            slot = 0;
+            block = getBlock(startOffset);
+            nextNumber();
+        }
+
+        @Override
+        boolean hasNext() {
+            return nextNumber != 0;
+        }
+
+        @Override
+        long next() {
+            if (hasNext()) {
+                long num = nextNumber;
+                try {
+                    nextNumber();
+                } catch (IOException ex) {
+                    ex.printStackTrace();
+                    nextNumber = 0;
+                }
+                return num;
+            }
+            throw new NoSuchElementException();
+        }
+
+        private void nextNumber() throws IOException {
+            if (slot < NUMBERS_IN_BLOCK) {
+                long nextNum = readNumber(block,slot++);
+                if (nextNum == 0) {     // end of the block, move to next one
+                    nextBlock();
+                } else {
+                    nextNumber = nextNum;
+                }
+            } else {
+               nextBlock();
+            }
+        }
+
+        private void nextBlock() throws IOException {
+            long nextBlock = getOffsetToNextBlock(block);
+
+            if (nextBlock == 0) { // end of list
+                nextNumber = 0;
+                return;
+            }
+            block = getBlock(nextBlock);
+            slot = 0;
+            nextNumber();
+        }
+    }
+
     @SuppressWarnings("serial")
-    private class BlockLRUCache<K, V> extends LinkedHashMap<K, V> {
-
+	private class BlockLRUCache extends LinkedHashMap<Long, byte[]> {
+        
         private static final int MAX_CAPACITY = 10000;
-
+        
         private BlockLRUCache() {
             super(MAX_CAPACITY,0.75f,true);
         }
 
-        protected boolean removeEldestEntry(Map.Entry<K, V> eldest) {
+        protected boolean removeEldestEntry(Map.Entry<Long, byte[]> eldest) {
             if (size()>MAX_CAPACITY) {
                 Object key = eldest.getKey();
                 if (!dirtyBlocks.contains(key)) {
